pyproject.toml
[tool.poetry]
name = "book-ocr"
version = "0.1.0"
description = ""
authors = ["Andy Garbett <andy@thinkactivelabs.co.uk>"]
readme = "README.md"
packages = [{include = "book_ocr"}]

[tool.poetry.dependencies]
python = "^3.11"
pytesseract = "^0.3.10"
python-dotenv = "^1.0.0"
opencv-python-headless = "^4.9.0.80"
numpy = "^1.26.3"
python-doctr = {extras = ["torch"], version = "^0.7.0"}
scikit-learn = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
